Refactored into separate classes
Logic placed in different classes to follow SOLID coding principles
